changed from `os.system` to `subprocess.call` in tools/rdp.py
for some reason `samples_dictionary(p)` was not finding the data file.
it seemed to be running before the RDP classifier is done saving the file.
adding a `time.sleep(30)` in modules/rdp.py before calling `samples_dictionary(p) worked for a small FASTA file, but IMO the cleaner solution is using `subprocess.call` instead of `os.system` in tools/rdp.py

subprocess was added in python 2.4, if backwards compat is an issue we can add a condition to check the version. this may require other changers as os.system was seeming to not block on my system, though all the docs I found said it would.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
# On branch wait-for-dict
# Changes to be committed:
#   (use "git reset HEAD <file>..." to unstage)
#
#	modified:   framework/modules/rdp.py
#	modified:   framework/tools/rdp.py
#