To make this work, I changed the shape of the taxa_color_dict returned by  that determines the colors in the piechart. Where previously it looked like , it now is 'flattened' to be like: {a_species:(color), a_phylum:(color) ... } with all the levels accessible through just the name of the specific taxon. This was changed in helper_functions.py and piechart.py is changed to expect it. I couldn't find anywhere else that used this result.
For an analysis at multiple levels, e.g. RDP, the colors work like this:
using the HSV color coordinate system, we take a circular slice at a given value, and split it up between all the taxons. Assuming that the levels are phylum, class, order, family, genus, first the program assigns each phylum a color  so that the hues are equally spaced for all the phyla, but all phyla have the same saturation and value. In other words, the colors are arranged at equal intervals on a circle of constant S and V. Now if we divide the circular slice into pie-slice sections, with each phylum's color in the middle of one, that slice is the range of hues available to members of that phylum. So then the classes in a specific phylum are given colors with hues evenly spaced within the hue-range of their phylum, with saturation increasing as we descend the tree.

Hopefully that makes some kind of sense. The result of it all is that for example, all the families in a given order or all the orders in a class will have colors with HSV hues similar to one another and to their parent, and genuses have higher saturation than families have higher saturation than orders.