Added ability to regard sequences as unclassified if they are given a low confidence by the RDP classifier.
From a user's perspective, there is now another field in the fasta upload form, "threshold" which they can specify a level below which the classifications will be ignored, and the sequence will be placed in a separate file they can downoad.

What the framework does is run RDP with the same options as normal, then ignore classifications below the threshold when building the samples dict. This allows it to say, use a confident classification at the order level but mark the family and genus as unclassified.